requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# GenAI and LLM
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
requests>=2.31.0

# Statistical analysis
statsmodels>=0.14.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Async support
asyncio-mqtt>=0.13.0

# File handling
python-docx>=0.8.11
PyPDF2>=3.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Optional: For enhanced functionality
# jupyter>=1.0.0
# ipykernel>=6.25.0